worker-based multithreading. shiny scene

+++ src/camera.rs
@@ -9,7 +9,7 @@ pub struct Camera {
     pub horizontal: Vec3,
     pub vertical: Vec3,
     pub lower_left: Vec3,
-    pub lens_radius: f32,
+    pub lens_radius: f64,
     pub u: Vec3,
     pub v: Vec3,
 }
@@ -17,7 +17,7 @@ pub struct Camera {
 impl Camera {
     fn random_in_unit_disk(rng: &mut ThreadRng) -> Vec3 {
         loop {
-            let p = 2.0 * Vec3(rng.gen::<f32>() - 0.5, rng.gen::<f32>() - 0.5, 0.0);
+            let p = 2.0 * Vec3(rng.gen::<f64>() - 0.5, rng.gen::<f64>() - 0.5, 0.0);
             if p.squared_length() < 1.0 {
                 return p;
             }
@@ -28,15 +28,15 @@ impl Camera {
         origin: Vec3,
         look_at: Vec3,
         up: Vec3,
-        fov: f32,
-        aspect: f32,
-        aperture: f32,
-        focus: f32,
+        fov: f64,
+        aspect: f64,
+        aperture: f64,
+        focus: f64,
     ) -> Camera {
-        let theta = fov * std::f32::consts::PI / 180.0;
+        let theta = fov * std::f64::consts::PI / 180.0;
         let half_width = (0.5 * theta).tan();
         let half_height = half_width / aspect;
-        // let theta = vfov * std::f32::consts::PI / 180.0;
+        // let theta = vfov * std::f64::consts::PI / 180.0;
         // let half_height = (theta * 0.5).tan();
         // let half_width = aspect * half_height;
         let w = (origin - look_at).normalized();
@@ -58,7 +58,7 @@ impl Camera {
             v,
         }
     }
-    pub fn get_ray(&self, s: f32, t: f32, rng: &mut ThreadRng) -> Ray {
+    pub fn get_ray(&self, s: f64, t: f64, rng: &mut ThreadRng) -> Ray {
         let rd = self.lens_radius * Self::random_in_unit_disk(rng);
         let offset = self.u * rd.x() + self.v * rd.y();
         Ray {
