+++ src/materials.rs
@@ -9,9 +9,9 @@ fn random_in_unit_sphere(rng: &mut ThreadRng) -> Vec3 {
     let mut v = Vec3(1.0, 1.0, 1.0);
     while v.squared_length() > 1.0 {
         v = Vec3(
-            2.0 * rng.gen::<f32>() - 1.0,
-            2.0 * rng.gen::<f32>() - 1.0,
-            2.0 * rng.gen::<f32>() - 1.0,
+            2.0 * rng.gen::<f64>() - 1.0,
+            2.0 * rng.gen::<f64>() - 1.0,
+            2.0 * rng.gen::<f64>() - 1.0,
         );
     }
     v
@@ -43,7 +43,7 @@ impl Material for DiffuseMaterial {
 
 pub struct MetalMaterial {
     pub albedo: Vec3,
-    pub fuzz: f32,
+    pub fuzz: f64,
 }
 
 impl Material for MetalMaterial {
@@ -59,14 +59,14 @@ impl Material for MetalMaterial {
 
 pub struct GlassMaterial {
     pub albedo: Vec3,
-    pub ref_idx: f32,
+    pub ref_idx: f64,
 }
 
 impl Material for GlassMaterial {
     fn scatter(&self, ray: &Ray, hit_record: &HitRecord, rng: &mut ThreadRng) -> (Ray, Vec3) {
         let outward_normal: Vec3;
-        let ni_over_nt: f32;
-        let cosine: f32;
+        let ni_over_nt: f64;
+        let cosine: f64;
 
         if ray.dir.dot(hit_record.normal) > 0.0 {
             outward_normal = -hit_record.normal;
@@ -78,7 +78,7 @@ impl Material for GlassMaterial {
             cosine = -ray.dir.dot(hit_record.normal) / ray.dir.length();
         }
 
-        if rng.gen::<f32>() > schlick(cosine, self.ref_idx) {
+        if rng.gen::<f64>() > schlick(cosine, self.ref_idx) {
             // borrowed this code from https://github.com/perliedman/raytracing-in-one-weekend/blob/master/src/material.rs
             match refract(&ray.dir, &outward_normal, ni_over_nt) {
                 Some(refraction) => {
@@ -110,7 +110,7 @@ fn reflect(v: &Vec3, n: &Vec3) -> Vec3 {
 }
 
 // borrowed this code from https://github.com/perliedman/raytracing-in-one-weekend/blob/master/src/material.rs
-fn refract(v: &Vec3, n: &Vec3, ni_over_nt: f32) -> Option<Vec3> {
+fn refract(v: &Vec3, n: &Vec3, ni_over_nt: f64) -> Option<Vec3> {
     let uv = v.normalized();
     let dt = uv.dot(*n);
     let discriminant = 1.0 - ni_over_nt * ni_over_nt * (1.0 - dt * dt);
@@ -122,7 +122,7 @@ fn refract(v: &Vec3, n: &Vec3, ni_over_nt: f32) -> Option<Vec3> {
 }
 
 // borrowed this code from https://github.com/perliedman/raytracing-in-one-weekend/blob/master/src/material.rs
-fn schlick(cosine: f32, ref_idx: f32) -> f32 {
+fn schlick(cosine: f64, ref_idx: f64) -> f64 {
     let r0 = (1.0 - ref_idx) / (1.0 + ref_idx);
     let r0sq = r0 * r0;
     r0sq + (1.0 - r0sq) * (1.0 - cosine).powf(5.0)
