+++ src/ray.rs
@@ -6,7 +6,7 @@ pub struct Ray {
 }
 
 impl Ray {
-    pub fn point_at_t(&self, t: f32) -> Vec3 {
+    pub fn point_at_t(&self, t: f64) -> Vec3 {
         self.pos + self.dir * t
     }
 }
